Hadith_app_Challenge-3/hadith_app/pyproject.toml
[project]
name = "hadith-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "gtts>=2.5.4",
    "python-dotenv>=1.1.0",
    "streamlit>=1.45.1",
    "stripe>=12.1.0",
    "supabase>=2.15.1",
]
